Added to settings explorer link and tari wallet version
